fix: int64 migration batch 5 - ILKernel and DefaultEngine fixes
INT64 Developer Guide compliance for recent rebase commits:

ILKernelGenerator fixes:
- Reduction.Arg.cs: All ArgMax/ArgMin helpers return long, take long totalSize
- Reduction.Axis.Simd.cs: Size/stride params, loop counters, arrays -> long
- Reduction.Axis.cs: Loop counters -> long
- Reduction.NaN.cs: Fix sealed -> static (rebase conflict)
- Masking.cs: Shape parameters -> long[]

DefaultEngine fixes:
- Default.Reduction.Nan.cs: All size/offset/stride/arrays -> long
- Default.NonZero.cs: CountNonZero returns long, arrays -> long[]
- Default.BooleanMask.cs: Size/count variables -> long

API fixes:
- TensorEngine.cs: CountNonZero returns long, remove duplicates
- np.count_nonzero.cs: Returns long
- np.any.cs, np.all.cs: Shape arrays -> long[], loop vars -> long

Rebase conflict fixes:
- np.random.rand.cs: Remove duplicate random() methods
- Delete duplicate Default.Op.Boolean.template.cs

Remaining: ~96 errors in Shape.Broadcasting.cs, NDArray`1.cs
See docs/INT64_MIGRATION_PROGRESS.md for full details

+++ src/NumSharp.Core/APIs/np.count_nonzero.cs
@@ -11,7 +11,7 @@ namespace NumSharp
         /// <param name="a">The array for which to count non-zeros.</param>
         /// <returns>Number of non-zero values in the array.</returns>
         /// <remarks>https://numpy.org/doc/stable/reference/generated/numpy.count_nonzero.html</remarks>
-        public static int count_nonzero(NDArray a)
+        public static long count_nonzero(NDArray a)
         {
             if (a.size == 0)
                 return 0;
@@ -42,7 +42,7 @@ namespace NumSharp
                 var result = np.zeros(new Shape(resultShape), NPTypeCode.Int64);
                 if (keepdims)
                 {
-                    var ks = new int[a.ndim];
+                    var ks = new long[a.ndim];
                     for (int d = 0, sd = 0; d < a.ndim; d++)
                         ks[d] = (d == axis) ? 1 : resultShape[sd++];
                     result.Storage.Reshape(new Shape(ks));
